refactor(sandbox): constants/theme.ts を theme/legacyColors.ts へ統合
1ファイルだけのレイヤーディレクトリ constants/ を解消し、実態
（テーマ移行中の旧カラー API）に合わせて theme/legacyColors.ts へ移動する。

- どこからも import されていない named re-export
  （ColorScheme / Spacing / SpacingName / Fonts）を削除
- ColorScheme は参照元 (ThemeContext / useTheme) が tokens/colors から
  直接 import するよう変更（barrel 的な再エクスポートを避ける運用に沿う）
- 互換 API（useTheme().colors）の廃止は既存のテーマ移行タスクの領分の
  ため踏み込まない
