+++ apps/sandbox/src/theme/ThemeContext.tsx
@@ -2,8 +2,8 @@ import type { ReactNode } from "react";
 import React, { createContext, useCallback, useContext, useEffect, useMemo, useState } from "react";
 import { Appearance, useColorScheme } from "react-native";
 import Storage from "expo-sqlite/kv-store";
-import { Colors, type ColorScheme, type ColorTokens } from "../constants/theme";
-import { semanticColors, type SemanticColorTokens } from "./tokens/colors";
+import { Colors, type ColorTokens } from "./legacyColors";
+import { type ColorScheme, semanticColors, type SemanticColorTokens } from "./tokens/colors";
 import { Radius } from "./tokens/radius";
 import { shadowsByScheme, type ShadowTokens } from "./tokens/shadows";
 import { Spacing } from "./tokens/spacing";
